changed half length to int so range() works

+++ Lab3/Palandrome.py
@@ -1,7 +1,7 @@
 
 testWord = raw_input("enter a word to check:")
 length = len(testWord)
-halfLength = length/2
+halfLength = int(length/2)
 #for loop checks the first character against the last,
 #then moves in a step and checks the next two sets,
 #because half of the length is an int there are no problems with odd numbers
